feat: add video playlist

+++ app_video/tasks.py
@@ -1,5 +1,5 @@
-import subprocess
 import os
+import time
 import shutil
 import logging
 from django.core.files.base import ContentFile
@@ -14,8 +14,12 @@ logger = logging.getLogger(__name__)
 
 @job
 def convert_video(instance_id, resolution_key):
+    time.sleep(1)
     try:
-        instance = Video.objects.get(pk=instance_id)
+        instance = Video.objects.filter(pk=instance_id).first()
+        if not instance:
+            logger.warning(f"Video {instance_id} nicht gefunden. Task abgebrochen.")
+            return
         ffmpeg_resolution = VIDEO_RESOLUTIONS.get(resolution_key)
 
         if not ffmpeg_resolution:
@@ -23,23 +27,22 @@ def convert_video(instance_id, resolution_key):
 
         source, target = get_video_paths(instance, resolution_key)
 
+
         cmd = [
             "ffmpeg",
-            "-i",
-            source,
-            "-s",
-            ffmpeg_resolution,
-            "-c:v",
-            "libx264",
-            "-crf",
-            "23",
-            "-c:a",
-            "aac",
+            "-i", source,
+            "-s", ffmpeg_resolution,
+            "-c:v", "libx264",
+            "-crf", "23",
+            "-c:a", "aac",
+            "-hls_time", "10",      
+            "-hls_list_size", "0",   
+            "-f", "hls",             
             "-y",
-            target,
+            target
         ]
 
-        if run_ffmpeg(cmd, f"Convert {resolution_key}"):
+        if run_ffmpeg(cmd, f"Convert {resolution_key} to HLS"):
             update_video_instance(instance, f"video_{resolution_key}", target)
 
     except Video.DoesNotExist:
@@ -48,43 +51,43 @@ def convert_video(instance_id, resolution_key):
 
 @job
 def create_thumbnail(instance_id):
+    time.sleep(1)
     try:
-        instance = Video.objects.get(pk=instance_id)
+        instance = Video.objects.filter(pk=instance_id).first()
+        if not instance:
+            logger.warning(f"Task abgebrochen: Video {instance_id} existiert nicht (mehr).")
+            return
         source, thumb_path = get_video_paths(instance)
 
-        cmd = [
-            "ffmpeg",
-            "-ss",
-            "00:00:01",
-            "-i",
-            source,
-            "-vframes",
-            "1",
-            "-q:v",
-            "2",
-            "-y",
-            thumb_path,
-        ]
+        logger.info(f"Source: {source}")
+        logger.info(f"Target Thumb: {thumb_path}")
 
         if not os.path.exists(source):
             logger.error(f"Quelldatei nicht gefunden: {source}")
             return
 
+        cmd = [
+            "ffmpeg", "-ss", "00:00:01", "-i", source,
+            "-vframes", "1", "-q:v", "2", "-y", thumb_path,
+        ]
+
         if run_ffmpeg(cmd, "Thumbnail"):
-            with open(thumb_path, "rb") as f:
-                instance.thumbnail_url.save(
-                    f"thumb_{instance.id}.jpg", ContentFile(f.read()), save=False
-                )
-                instance.save(update_fields=["thumbnail_url"])
-            os.remove(thumb_path)
-            logger.info(f"SUCCESS: Thumbnail für {instance.title} erstellt.")
+            if os.path.exists(thumb_path):
+                with open(thumb_path, "rb") as f:
+                    file_name = os.path.basename(thumb_path)
+                    instance.thumbnail_url.save(
+                        file_name,
+                        ContentFile(f.read()),
+                        save=True  
+                    )
+                os.remove(thumb_path)
+                logger.info(f"SUCCESS: Thumbnail für {instance.title} erstellt.")
+            else:
+                logger.error(f"FFmpeg war erfolgreich, aber Datei fehlt unter: {thumb_path}")
 
-    except Video.DoesNotExist:
-        logger.error(f"Video with ID {instance_id} not found.")
     except Exception as e:
-        logger.error(f"An unexpected error occurred during thumbnail creation: {e}")
-
-
+        logger.error(f"Fehler beim Thumbnail: {e}", exc_info=True)
+        
 def delete_video_directory(instance):
     """
     Deletes the entire directory containing the video,
